+++ packages/forms/README.md
@@ -33,19 +33,19 @@ These features require the following minimum versions of the dependencies:
 import '@zendeskgarden/react-forms/dist/styles.css';
 
 import { ThemeProvider } from '@zendeskgarden/react-theming';
-import { FieldProvider, Label, Hint, Input, Message } from '@zendeskgarden/react-forms';
+import { Field, Label, Hint, Input, Message } from '@zendeskgarden/react-forms';
 
 /**
  * Place a `ThemeProvider` at the root of your React application
  */
 <ThemeProvider>
   <form>
-    <FieldProvider>
+    <Field>
       <Label>Example Text Input</Label>
       <Hint>Hint text</Hint>
       <Input placeholder="Accepts all native input props" />
       <Message>Default message styling</Message>
-    </FieldProvider>
+    </Field>
   </form>
 </ThemeProvider>;
 ```
