+++ packages/forms/examples/media-input.md
@@ -2,8 +2,8 @@
 const SearchIcon = require('@zendeskgarden/svg-icons/src/16/search-stroke.svg').default;
 const ShieldIcon = require('@zendeskgarden/svg-icons/src/16/shield-stroke.svg').default;
 
-<FieldProvider>
+<Field>
   <Label>Test Label</Label>
   <MediaInput start={<SearchIcon />} end={<ShieldIcon />} />
-</FieldProvider>;
+</Field>;
 ```
