+++ packages/forms/src/fields/Toggle.js
@@ -9,7 +9,7 @@ import React, { createContext } from 'react';
 import PropTypes from 'prop-types';
 import { useKeyboardFocus } from '@zendeskgarden/container-keyboardfocus';
 import { composeEventHandlers } from '@zendeskgarden/container-utilities';
-import { FieldContext } from './common/FieldProvider';
+import { FieldContext } from './common/Field';
 import useFieldContext from '../utils/useFieldContext';
 import { StyledToggle, StyledToggleInput } from '../styled';
