+++ packages/forms/src/fields/common/Message.spec.js
@@ -7,14 +7,14 @@
 
 import React from 'react';
 import { render } from 'garden-test-utils';
-import { FieldProvider, Checkbox, Radio, Toggle, Message } from '../../';
+import { Field, Checkbox, Radio, Toggle, Message } from '../../';
 
 describe('Hint', () => {
   it('renders text label by default', () => {
     const { getByTestId } = render(
-      <FieldProvider>
+      <Field>
         <Message data-test-id="message">Test</Message>
-      </FieldProvider>
+      </Field>
     );
 
     expect(getByTestId('message')).toHaveAttribute('data-garden-id', 'forms.text_message');
@@ -22,11 +22,11 @@ describe('Hint', () => {
 
   it('renders checkbox message if within a Checkbox component', () => {
     const { getByTestId } = render(
-      <FieldProvider>
+      <Field>
         <Checkbox>
           <Message data-test-id="message">Test</Message>
         </Checkbox>
-      </FieldProvider>
+      </Field>
     );
 
     expect(getByTestId('message')).toHaveAttribute('data-garden-id', 'forms.checkbox_message');
@@ -34,11 +34,11 @@ describe('Hint', () => {
 
   it('renders toggle message if within a Toggle component', () => {
     const { getByTestId } = render(
-      <FieldProvider>
+      <Field>
         <Toggle>
           <Message data-test-id="message">Test</Message>
         </Toggle>
-      </FieldProvider>
+      </Field>
     );
 
     expect(getByTestId('message')).toHaveAttribute('data-garden-id', 'forms.toggle_message');
@@ -46,11 +46,11 @@ describe('Hint', () => {
 
   it('renders radio message if within a Radio component', () => {
     const { getByTestId } = render(
-      <FieldProvider>
+      <Field>
         <Radio>
           <Message data-test-id="message">Test</Message>
         </Radio>
-      </FieldProvider>
+      </Field>
     );
 
     expect(getByTestId('message')).toHaveAttribute('data-garden-id', 'forms.radio_message');
